refactor: migrate scripts/visual_qa_zones.py to src/quality/ (#344)
T8 of ADR-0010 implementation. Moves visual_qa_zones to src/quality/
and updates caller:
- scripts/economist_agent.py (lazy import at line 515)

Refs #344
